dev-requirements.txt
parameterized
pytest
pytest-cov
torchsnapshot-nightly
pyre-check
torchvision
